fix: fix authforminput focus issue and implement autocomplete attributes
