+++ src/app/auth/(sign)/signin/page.tsx
@@ -1,6 +1,21 @@
 'use client';
-import AuthPage from '@/components/auth/AuthPage';
+import AuthTitle from '@/components/auth/AuthTitle';
+import Divider from '@/components/auth/Divider';
+import GoogleSignUpButton from '@/components/auth/GoogleButton';
+import SignInForm from '@/components/auth/signin/SignInForm';
+import useSignInForm from '@/hooks/auth/useSignInForm';
 
-export default function Page() {
-  return <AuthPage type="signin" />;
+export default function AuthPage() {
+  const signInFormProps = useSignInForm();
+
+  return (
+    <div className="flex h-full flex-col items-center justify-center px-5 pb-8 sm:pb-0">
+      <div className="flex w-full max-w-[400px] flex-col gap-6">
+        <AuthTitle>登入 Talents 帳戶</AuthTitle>
+        <SignInForm {...signInFormProps} />
+        <Divider>或</Divider>
+        <GoogleSignUpButton isSubmitting={signInFormProps.isSubmitting} />
+      </div>
+    </div>
+  );
 }
